pyproject.toml
[tool.poetry]
name = "\"ai-voice-chatbot\""
version = "0.1.0"
description = "The chatbot is for healthcare telemedicine services\""
authors = ["Bless50 <blessn691@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
pyttsx3 = "^2.97"
python-dotenv = "^1.0.1"
llama-index-llms-groq = "^0.2.0"
groq = "^0.11.0"
pyaudio = "^0.2.14"
llama-index-readers-file = "^0.2.2"
llama-index-embeddings-gemini = "^0.2.0"
llama-index = "^0.11.13"
google-generativeai = ">=0.3.0"
matplotlib = "^3.9.2"
llama-index-core = "^0.11.14"
flask-socketio = "^5.3.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
